Id 12 admin update role model

+++ backend/src/controllers/roleModelController.ts
@@ -2,8 +2,9 @@ import { Request, Response } from "express";
 import {
   deleteRoleModelById,
   getAllRoleModels,
+  addRoleModel,
+  updateRoleModel,
 } from "../services/roleModelService";
-import { addRoleModel } from "../services/roleModelService";
 import { IRoleModel } from "../models/interfaces";
 
 /**
@@ -31,7 +32,7 @@ export const addRoleModels = async (req: Request, res: Response) => {
   try {
     const newRoleModel = await addRoleModel(roleModelData);
     if (!newRoleModel) {
-      return res.status(500).json({ message: "Failed to add role model" });
+      return res.status(404).json({ message: "Failed to find role model" });
     }
     return res.status(201).json(newRoleModel);
   } catch (error) {
@@ -43,7 +44,11 @@ export const addRoleModels = async (req: Request, res: Response) => {
  * Controller to delete a role model
  */
 export const deleteRoleModel = async (req: Request, res: Response) => {
-  const roleModelId: string = req.params.id;
+  const { roleModelId } = req.params;
+
+  if (!roleModelId) {
+    return res.status(400).json({ message: "Role model ID is required" });
+  }
 
   try {
     const result = await deleteRoleModelById(roleModelId);
@@ -53,6 +58,28 @@ export const deleteRoleModel = async (req: Request, res: Response) => {
         .json({ message: "Role model not found with ID: ", roleModelId });
     }
     return res.status(204).send();
+  } catch (error) {
+    return res.status(500).json({ message: "Server error", error });
+  }
+};
+
+/**
+ * Controller to update a role model
+ */
+export const putRoleModels = async (req: Request, res: Response) => {
+  const roleModelData: IRoleModel = req.body;
+  const { roleModelId } = req.params;
+
+  if (!roleModelId) {
+    return res.status(400).json({ message: "Role model ID is required" });
+  }
+
+  try {
+    const updatedRoleModel = await updateRoleModel(roleModelId, roleModelData);
+    if (!updatedRoleModel) {
+      return res.status(404).json({ message: "Failed to find role model" });
+    }
+    return res.status(200).json(updatedRoleModel);
   } catch (error) {
     return res.status(500).json({ message: "Server error", error });
   }
